pyproject.toml
[tool.poetry]
name = "divvundev"
version = "0.1.0"
description = "Configure, build and test giellalt repos"
authors = ["Flammie A Pirinen <flammie@iki.fi>"]
readme = "README.md"
homepage = "https://divvun.github.io/divvundev"
repository = "https://github.com/divvun/divvundev"
license = "GPL-3.0"

[tool.poetry.scripts]
divvundev = "divvundev.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
target-version = "py39"

[tool.ruff.lint]
select = [
    "F", #Pyflakes
    "B", #flake8-bugbear
    "A", #flake8-builtins 
    "C4", #flake8-comprehensions
    "ERA", #eradicate
    "I", #isort
    "PT", #flake8-pytest-style
    "C90", #mccabe
    "N", #pep8-naming
    "E", #pycodestyle error
    "PL" #Pylint
]
ignore = ["A003"]

[tool.poetry.dependencies]
python = "^3.9"
gitpython = "^3.1.44"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.2"

